refactor: separate functions
