+++ frontend/rust-lib/flowy-core/src/deps_resolve/collab_deps.rs
@@ -11,10 +11,10 @@ use flowy_sqlite::{
   prelude::*,
   schema::{collab_snapshot, collab_snapshot::dsl},
 };
-use flowy_user::services::UserSession;
+use flowy_user::services::UserManager;
 use lib_infra::util::timestamp;
 
-pub struct SnapshotDBImpl(pub Weak<UserSession>);
+pub struct SnapshotDBImpl(pub Weak<UserManager>);
 
 impl SnapshotPersistence for SnapshotDBImpl {
   fn get_snapshots(&self, uid: i64, object_id: &str) -> Vec<CollabSnapshot> {
