+++ frontend/rust-lib/flowy-core/src/module.rs
@@ -3,13 +3,13 @@ use std::sync::Weak;
 use flowy_database2::DatabaseManager;
 use flowy_document2::manager::DocumentManager as DocumentManager2;
 use flowy_folder2::manager::FolderManager;
-use flowy_user::services::UserSession;
+use flowy_user::services::UserManager;
 use lib_dispatch::prelude::AFPlugin;
 
 pub fn make_plugins(
   folder_manager: Weak<FolderManager>,
   database_manager: Weak<DatabaseManager>,
-  user_session: Weak<UserSession>,
+  user_session: Weak<UserManager>,
   document_manager2: Weak<DocumentManager2>,
 ) -> Vec<AFPlugin> {
   let store_preferences = user_session
