+++ frontend/rust-lib/flowy-user/src/services/mod.rs
@@ -1,7 +1,10 @@
-pub use user_session::*;
+pub use manager::*;
 
 pub mod database;
-pub mod session_serde;
-mod user_session;
+pub mod entities;
+mod historical_user;
+mod manager;
+mod user_awareness;
 mod user_sql;
+mod user_workspace;
 mod user_workspace_sql;
